// utils/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  JSON: any;
  bigint: any;
  bytea: any;
  citext: any;
  jsonb: any;
  timestamptz: any;
  uuid: any;
};
// ...
export type StripeAddress = {
  __typename?: 'StripeAddress';
  /** City/Ward. */
  city?: Maybe<Scalars['String']>;
  /** Two-letter country code ([ISO 3166-1 alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2)). */
  country?: Maybe<Scalars['String']>;
  /** Block/Building number. */
  line1?: Maybe<Scalars['String']>;
  /** Building details. */
  line2?: Maybe<Scalars['String']>;
  /** ZIP or postal code. */
  postalCode?: Maybe<Scalars['String']>;
  /** Prefecture. */
  state?: Maybe<Scalars['String']>;
};
// ...
export type StripeCharges = {
  __typename?: 'StripeCharges';
  data: Array<StripeCharge>;
  hasMore: Scalars['Boolean'];
  object: Scalars['String'];
  url: Scalars['String'];
};
// ...
export type StripeCustomer = {
  __typename?: 'StripeCustomer';
  /** The customer's address. */
  address?: Maybe<StripeAddress>;
  /** Current balance, if any, being stored on the customer. If negative, the customer has credit to apply to their next invoice. If positive, the customer has an amount owed that will be added to their next invoice. The balance does not refer to any unpaid invoices; it solely takes into account amounts that have yet to be successfully applied to any invoice. This balance is only taken into account as invoices are finalized. */
  balance: Scalars['Int'];
  charges: StripeCharges;
  /** Time at which the object was created. Measured in seconds since the Unix epoch. */
  created: Scalars['Int'];
  /** Three-letter [ISO code for the currency](https://stripe.com/docs/currencies) the customer can be charged in for recurring billing purposes. */
  currency?: Maybe<Scalars['String']>;
  /**
   * When the customer's latest invoice is billed by charging automatically, `delinquent` is `true` if the invoice's latest charge failed. When the customer's latest invoice is billed by sending an invoice, `delinquent` is `true` if the invoice isn't paid by its due date.
   *
   * If an invoice is marked uncollectible by [dunning](https://stripe.com/docs/billing/automatic-collection), `delinquent` doesn't get reset to `false`.
   */
  delinquent?: Maybe<Scalars['Boolean']>;
  /** An arbitrary string attached to the object. Often useful for displaying to users. */
  description?: Maybe<Scalars['String']>;
  /** The customer's email address. */
  email?: Maybe<Scalars['String']>;
  /** Unique identifier for the object. */
  id: Scalars['String'];
  /** The prefix for the customer used to generate unique invoice numbers. */
  invoicePrefix?: Maybe<Scalars['String']>;
  invoices: StripeInvoices;
  /** Has the value `true` if the object exists in live mode or the value `false` if the object exists in test mode. */
  livemode: Scalars['Boolean'];
  /** Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. */
  metadata: Scalars['JSON'];
  /** The customer's full name or business name. */
  name?: Maybe<Scalars['String']>;
  /** The suffix of the customer's next invoice number, e.g., 0001. */
  nextInvoiceSequence?: Maybe<Scalars['Int']>;
  /** String representing the object's type. Objects of the same type share the same value.  */
  object: Scalars['String'];
  paymentIntents: StripePaymentIntents;
  paymentMethods: StripePaymentMethods;
  /** The customer's phone number. */
  phone?: Maybe<Scalars['String']>;
  /** The customer's preferred locales (languages), ordered by preference. */
  preferredLocales?: Maybe<Array<Scalars['String']>>;
  /** Mailing and shipping address for the customer. Appears on invoices emailed to this customer. */
  shipping?: Maybe<StripeCustomerShipping>;
  /** The customer's current subscriptions, if any. */
  subscriptions: StripeSubscriptions;
  tax?: Maybe<StripeCustomerTax>;
};
// ...
export type StripeInvoices = {
  __typename?: 'StripeInvoices';
  data: Array<StripeInvoice>;
  /** True if this list has another page of items after this one that can be fetched. */
  hasMore: Scalars['Boolean'];
  object: Scalars['String'];
  /** The URL where this list can be accessed. */
  url: Scalars['String'];
};
// ...
export type StripePaymentIntents = {
  __typename?: 'StripePaymentIntents';
  data: Array<StripePaymentIntent>;
  hasMore: Scalars['Boolean'];
  object: Scalars['String'];
  url: Scalars['String'];
};
// ...
export type StripePaymentMethods = {
  __typename?: 'StripePaymentMethods';
  data: Array<StripePaymentMethod>;
  /** True if this list has another page of items after this one that can be fetched. */
  hasMore: Scalars['Boolean'];
  object: Scalars['String'];
  /** The URL where this list can be accessed. */
  url: Scalars['String'];
};
// ...
export type StripePrice = {
  __typename?: 'StripePrice';
  /** Whether the price can be used for new purchases. */
  active: Scalars['Boolean'];
  /** Describes how to compute the price per period. Either `per_unit` or `tiered`. `per_unit` indicates that the fixed amount (specified in `unit_amount` or `unit_amount_decimal`) will be charged per unit in `quantity` (for prices with `usage_type=licensed`), or per unit of total usage (for prices with `usage_type=metered`). `tiered` indicates that the unit pricing will be computed using a tiering strategy as defined using the `tiers` and `tiers_mode` attributes. */
  billingScheme: Scalars['String'];
  /** Time at which the object was created. Measured in seconds since the Unix epoch. */
  created: Scalars['Int'];
  /** Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies). */
  currency: Scalars['String'];
  /** Unique identifier for the object. */
  id: Scalars['String'];
  /** Has the value `true` if the object exists in live mode or the value `false` if the object exists in test mode. */
  livemode: Scalars['Boolean'];
  /** A lookup key used to retrieve prices dynamically from a static string. This may be up to 200 characters. */
  lookupKey?: Maybe<Scalars['String']>;
  /** Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. */
  metadata?: Maybe<Scalars['JSON']>;
  /** A brief description of the price, hidden from customers. */
  nickname?: Maybe<Scalars['String']>;
  /** String representing the object's type. Objects of the same type share the same value. */
  object: Scalars['String'];
  /** The ID of the product this price is associated with. */
  product: StripeProduct;
  /** Defines if the tiering price should be `graduated` or `volume` based. In `volume`-based tiering, the maximum quantity within a period determines the per unit price. In `graduated` tiering, pricing can change as the quantity grows. */
  tiersMode?: Maybe<Scalars['String']>;
  /** One of `one_time` or `recurring` depending on whether the price is for a one-time purchase or a recurring (subscription) purchase. */
  type: Scalars['String'];
  /** The unit amount in %s to be charged, represented as a whole integer if possible. Only set if `billing_scheme=per_unit`. */
  unitAmount?: Maybe<Scalars['Int']>;
  /** The unit amount in %s to be charged, represented as a decimal string with at most 12 decimal places. Only set if `billing_scheme=per_unit`. */
  unitAmountDecimal?: Maybe<Scalars['String']>;
};
// ...
export type StripeSubscriptionItem = {
  __typename?: 'StripeSubscriptionItem';
  /** Define thresholds at which an invoice will be sent, and the related subscription advanced to a new billing period */
  billingThresholds?: Maybe<StripeSubscriptionItemBillingThresholds>;
  /** Time at which the object was created. Measured in seconds since the Unix epoch. */
  created: Scalars['Int'];
  /** Unique identifier for the object. */
  id: Scalars['String'];
  /** Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. */
  metadata: Scalars['JSON'];
  /** String representing the object's type. Objects of the same type share the same value. */
  object: Scalars['String'];
  /**
   * You can now model subscriptions more flexibly using the [Prices API](https://stripe.com/docs/api#prices). It replaces the Plans API and is backwards compatible to simplify your migration.
   *
   * Plans define the base price, currency, and billing cycle for recurring purchases of products.
   *
   * [Products](https://stripe.com/docs/api#products) help you track inventory or provisioning, and plans help you track pricing. Different physical goods or levels of service should be represented by products, and pricing options should be represented by plans. This approach lets you change prices without having to change your provisioning scheme.
   *
   * For example, you might have a single "gold" product that has plans for $10/month, $100/year, €9/month, and €90/year.
   *
   * Related guides: [Set up a subscription](https://stripe.com/docs/billing/subscriptions/set-up-subscription) and more about [products and prices](https://stripe.com/docs/products-prices/overview).
   */
  plan: StripePlan;
  /**
   * Prices define the unit cost, currency, and (optional) billing cycle for both recurring and one-time purchases of products.
   *
   * [Products](https://stripe.com/docs/api#products) help you track inventory or provisioning, and prices help you track payment terms. Different physical goods or levels of service should be represented by products, and pricing options should be represented by prices. This approach lets you change prices without having to change your provisioning scheme.
   *
   * For example, you might have a single "gold" product that has prices for $10/month, $100/year, and €9 once.
   *
   * Related guides: [Set up a subscription](https://stripe.com/docs/billing/subscriptions/set-up-subscription), [create an invoice](https://stripe.com/docs/billing/invoices/create), and more about [products and prices](https://stripe.com/docs/products-prices/overview).
   */
  price: StripePrice;
  /** The [quantity](https://stripe.com/docs/subscriptions/quantities) of the plan to which the customer should be subscribed. */
  quantity?: Maybe<Scalars['Int']>;
  /** The `subscription` this `subscription_item` belongs to. */
  subscription: Scalars['String'];
};

export type StripeSubscriptionItemBillingThresholds = {
  __typename?: 'StripeSubscriptionItemBillingThresholds';
  /** Usage threshold that triggers the subscription to create an invoice */
  usageGte?: Maybe<Scalars['Int']>;
};

export type StripeSubscriptionItems = {
  __typename?: 'StripeSubscriptionItems';
  data: Array<StripeSubscriptionItem>;
  /** True if this list has another page of items after this one that can be fetched. */
  hasMore: Scalars['Boolean'];
  object: Scalars['String'];
  /** The URL where this list can be accessed. */
  url: Scalars['String'];
};

export type StripeSubscriptionPauseCollection = {
  __typename?: 'StripeSubscriptionPauseCollection';
  /** The payment collection behavior for this subscription while paused. One of `keep_as_draft`, `mark_uncollectible`, or `void`. */
  behavior: Scalars['String'];
  /** The time after which the subscription will resume collecting payments. */
  resumesAt?: Maybe<Scalars['Int']>;
};

export type StripeSubscriptions = {
  __typename?: 'StripeSubscriptions';
  data: Array<StripeSubscription>;
  /** True if this list has another page of items after this one that can be fetched. */
  hasMore: Scalars['Boolean'];
  object: Scalars['String'];
  /** The URL where this list can be accessed. */
  url: Scalars['String'];
};
// ...
export type PlanFragment = { __typename?: 'plans', id: any, name: string, description?: string | null, amount: any, currency: string, stripePriceId: string };

export type UserWithSubscriptionFragment = { __typename?: 'users', id: any, profile?: { __typename?: 'profiles', id: any, stripeCustomer: { __typename?: 'StripeCustomer', subscriptions: { __typename?: 'StripeSubscriptions', data: Array<{ __typename?: 'StripeSubscription', id: string, items: { __typename?: 'StripeSubscriptionItems', data: Array<{ __typename?: 'StripeSubscriptionItem', id: string, price: { __typename?: 'StripePrice', id: string } }> } }> } } } | null };